% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costs.R
\name{mcb_categories}
\alias{mcb_categories}
\title{Commodity categories}
\usage{
mcb_categories()
}
\value{
A character vector of the commodity category labels.
}
\description{
Commodity categories
}
\examples{
mcb_categories()
}
