% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costs.R
\name{mcb_commodities}
\alias{mcb_commodities}
\title{List commodities}
\usage{
mcb_commodities(category = NULL)
}
\arguments{
\item{category}{Optional character vector of categories to filter by
(see \code{\link{mcb_categories}}).}
}
\value{
A character vector of commodity names.
}
\description{
List commodities
}
\examples{
mcb_commodities()
mcb_commodities("Precious Metals")
}
