% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drivers.R
\name{mcb_cross_cutting}
\alias{mcb_cross_cutting}
\title{Cross-cutting cost drivers}
\usage{
mcb_cross_cutting(category = NULL)
}
\arguments{
\item{category}{Optional character vector to filter \code{driver_category}.}
}
\value{
A data frame of cross-cutting cost drivers.
}
\description{
Cross-cutting cost drivers
}
\examples{
mcb_cross_cutting()
mcb_cross_cutting("Energy")
}
