% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{mcb_sources}
\alias{mcb_sources}
\title{Interpretation notes and indicative sources}
\usage{
mcb_sources(section = NULL)
}
\arguments{
\item{section}{Optional: "NOTES" or "SOURCES" to filter.}
}
\value{
A data frame of notes/sources.
}
\description{
Interpretation notes and indicative sources
}
\examples{
mcb_sources("SOURCES")
mcb_sources("NOTES")
}
