% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy.R
\name{mcb_strategic_groups}
\alias{mcb_strategic_groups}
\title{Strategic thematic groups}
\usage{
mcb_strategic_groups()
}
\value{
A character vector of the four thematic groups.
}
\description{
Strategic thematic groups
}
\examples{
mcb_strategic_groups()
}
