% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costs.R
\name{mcb_vintages}
\alias{mcb_vintages}
\title{Available cost vintages}
\usage{
mcb_vintages()
}
\value{
A character vector of the snapshot vintages present in the data.
}
\description{
Available cost vintages
}
\examples{
mcb_vintages()
}
